requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
pydantic
mysql-connector-python
python-dotenv
prefect
fastapi
uvicorn[standard]
python-dotenv
mysql-connector-python
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
email-validator
python-multipart
fastapi
uvicorn[standard]
python-dotenv
mysql-connector-python
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
email-validator
python-multipart
